Factor out pile to its own module
