+++ crossbeam-epoch/src/sync/mod.rs
@@ -4,4 +4,5 @@
 #[cfg(not(crossbeam_loom))]
 pub(crate) mod once_lock;
 
+pub(crate) mod pile;
 pub(crate) mod rwlock;
